frameAllGenes

This function creates a data.frame for each result of the diffrential expression method performed. It was created to ensure that all genes been listed, even thought the gene does not appear as diffrentially expressed, in this case NA data is inserted. This functis is executed into runExpression() function, so there is no need to call it separatelly.

Usage

frameAllGenes(cons_list, countMatrix)

Arguments

cons_list

A named list of data.frames, each corresponding to a method (e.g., edgeR, DESeq2, etc.)

countMatrix

RData object with table count, where line name is gene name, and name column is treat name

Value

list with analysis expression containing: all methods executed and their results, ensuring thar all genes has been listed;

Examples

data(gse95077)
treats <- c("BM", "JJ")
cons_result <- runExpression(
  numberReplics = 3,
  groupName = treats,
  rDataFrameCount = gse95077,
  experimentName = "test_cons",
  controlDeseq2 = "BM",
  contrastDeseq2 = "JJ",
  outDirPath = tempdir(),
  printResults = FALSE
)
#> calcNormFactors has been renamed to normLibSizes
#> Using classic mode.
#> Getting annotation of the Homo Sapiens...
#> Using reference genome 38.
#> Calculating gene expression values...
#> RQ fit ......
#> SQN 
#> Warning: KnowSeq failed and will be skipped: KnowSeq execution failed due to an internal dependency issue involving the 'mclust' package: could not find function "mclustBIC"
#> calcNormFactors has been renamed to normLibSizes
#> Removing intercept from test coefficients
#> [1] "Computing (M,D) values..."
#> [1] "Computing probability of differential expression..."
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Initial number of DE patterns = 2
#> Final number of DE patterns = 2
#> converting counts to integer mode
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> Estimating sequencing depths...
#> Resampling to get new data matrices...
#> perm= 1
#> perm= 2
#> perm= 3
#> perm= 4
#> perm= 5
#> perm= 6
#> perm= 7
#> perm= 8
#> perm= 9
#> perm= 10
#> perm= 11
#> perm= 12
#> perm= 13
#> perm= 14
#> perm= 15
#> perm= 16
#> perm= 17
#> perm= 18
#> perm= 19
#> perm= 20
#> perm= 21
#> perm= 22
#> perm= 23
#> perm= 24
#> perm= 25
#> perm= 26
#> perm= 27
#> perm= 28
#> perm= 29
#> perm= 30
#> perm= 31
#> perm= 32
#> perm= 33
#> perm= 34
#> perm= 35
#> perm= 36
#> perm= 37
#> perm= 38
#> perm= 39
#> perm= 40
#> perm= 41
#> perm= 42
#> perm= 43
#> perm= 44
#> perm= 45
#> perm= 46
#> perm= 47
#> perm= 48
#> perm= 49
#> perm= 50
#> perm= 51
#> perm= 52
#> perm= 53
#> perm= 54
#> perm= 55
#> perm= 56
#> perm= 57
#> perm= 58
#> perm= 59
#> perm= 60
#> perm= 61
#> perm= 62
#> perm= 63
#> perm= 64
#> perm= 65
#> perm= 66
#> perm= 67
#> perm= 68
#> perm= 69
#> perm= 70
#> perm= 71
#> perm= 72
#> perm= 73
#> perm= 74
#> perm= 75
#> perm= 76
#> perm= 77
#> perm= 78
#> perm= 79
#> perm= 80
#> perm= 81
#> perm= 82
#> perm= 83
#> perm= 84
#> perm= 85
#> perm= 86
#> perm= 87
#> perm= 88
#> perm= 89
#> perm= 90
#> perm= 91
#> perm= 92
#> perm= 93
#> perm= 94
#> perm= 95
#> perm= 96
#> perm= 97
#> perm= 98
#> perm= 99
#> perm= 100
#> Number of thresholds chosen (all possible thresholds) = 8
#> Getting all the cutoffs for the thresholds...
#> Getting number of false positives in the permutation...
#> Differential expression analysis completed.

genes_list <- results(cons_result)
framed <- frameAllGenes(cons_list = genes_list, countMatrix = gse95077)
#>  - 1edger
#>  - 2limma
#>  - 3noiseq
#>  - 4ebseq
#>  - 5deseq2
#>  - 6samseq
head(framed)
#> $edger
#>                       logFC   logCPM       PValue          FDR
#> ENSG00000008018 -0.79096798 13.41535 2.454040e-07 4.754703e-07
#> ENSG00000044574  0.76516594 14.70636 6.550355e-08 1.353740e-07
#> ENSG00000065978  1.37692287 14.54283 1.006684e-21 5.201200e-21
#> ENSG00000068697 -0.26297090 11.43314 9.196575e-02 1.239538e-01
#> ENSG00000075624  2.41956920 17.28991 1.592265e-64 1.645341e-63
#> ENSG00000100138 -0.15422039 14.22014 2.814168e-01 3.231082e-01
#> ENSG00000101361  0.84916553 15.04102 1.898568e-09 4.904634e-09
#> ENSG00000105193 -0.04863955 15.15030 7.298942e-01 7.542240e-01
#> ENSG00000105887 -2.58726793 15.49899 1.083279e-67 1.679082e-66
#> ENSG00000105968 -2.29178521 14.81767 2.047152e-53 1.269234e-52
#> ENSG00000107223  1.31199225 12.85509 2.289503e-18 1.013923e-17
#> ENSG00000107581 -0.86791803 15.43626 1.118705e-09 3.152714e-09
#> ENSG00000115758  1.00006665 15.46416 1.172654e-11 3.635228e-11
#> ENSG00000119421 -0.82123504 11.56161 5.181036e-08 1.147230e-07
#> ENSG00000127884  0.47421481 11.44648 2.202170e-03 3.413364e-03
#> ENSG00000131495  0.91643625 11.25444 2.162180e-09 5.155968e-09
#> ENSG00000138279  0.19563288 13.04901 1.956555e-01 2.426129e-01
#> ENSG00000142541  0.64758559 13.65936 9.105500e-06 1.485634e-05
#> ENSG00000142937  0.21337046 14.86305 1.310209e-01 1.692354e-01
#> ENSG00000143158  0.05935631 11.39980 6.931866e-01 7.409926e-01
#> ENSG00000147677 -0.34303930 14.88690 1.683271e-02 2.371883e-02
#> ENSG00000154723  0.69174956 12.88277 2.780728e-06 4.789031e-06
#> ENSG00000161016 -0.41146987 17.60368 2.746848e-03 4.054871e-03
#> ENSG00000163956  1.10557165 12.89310 1.121156e-13 3.861760e-13
#> ENSG00000165502  0.12734943 13.16697 3.872063e-01 4.286927e-01
#> ENSG00000166337  0.16702683 11.07777 2.730659e-01 3.231082e-01
#> ENSG00000166710 -0.01314995 17.28763 9.235572e-01 9.235572e-01
#> ENSG00000168264 -2.44851621 13.95057 8.905299e-58 6.901607e-57
#> ENSG00000175130  3.11050301 12.87483 5.838844e-85 1.810042e-83
#> ENSG00000178741 -0.81034968 12.83824 3.101324e-07 5.655356e-07
#> ENSG00000182117 -1.19131064 13.10237 9.949361e-16 3.855377e-15
#> 
#> $limma
#>                              ID       logFC  AveExpr           t      P.Value
#> ENSG00000008018 ENSG00000008018 -0.95196356 13.34118  -5.3613826 4.127330e-04
#> ENSG00000044574 ENSG00000044574  0.60752254 14.65482   8.4177759 1.224851e-05
#> ENSG00000065978 ENSG00000065978  1.21507377 14.38345  14.8198926 9.155662e-08
#> ENSG00000068697 ENSG00000068697 -0.39587110 11.40842  -2.4021929 3.901116e-02
#> ENSG00000075624 ENSG00000075624  2.26462593 16.83075  35.4283510 3.287295e-11
#> ENSG00000100138 ENSG00000100138 -0.30581829 14.21484  -4.3715490 1.672912e-03
#> ENSG00000101361 ENSG00000101361  0.69974471 14.97639   9.8791661 3.184299e-06
#> ENSG00000105193 ENSG00000105193 -0.20714808 15.14516  -3.0008062 1.446714e-02
#> ENSG00000105887 ENSG00000105887 -2.72757693 14.97685 -36.1720980 2.715081e-11
#> ENSG00000105968 ENSG00000105968 -2.43489510 14.39846 -27.3269592 3.568767e-10
#> ENSG00000107223 ENSG00000107223  1.13541252 12.70573  10.1709174 2.482665e-06
#> ENSG00000107581 ENSG00000107581 -1.00735588 15.36346 -13.2368737 2.497675e-07
#> ENSG00000115758 ENSG00000115758  0.88224303 15.35972   7.2593918 4.093172e-05
#> ENSG00000119421 ENSG00000119421 -0.97611546 11.50021  -7.5161882 3.095407e-05
#> ENSG00000127884 ENSG00000127884  0.31186589 11.41054   1.4322479 1.849320e-01
#> ENSG00000131495 ENSG00000131495  0.73606923 11.17528   4.0730938 2.622583e-03
#> ENSG00000138279 ENSG00000138279  0.03202230 13.03027   0.1800196 8.610128e-01
#> ENSG00000142541 ENSG00000142541  0.49330254 13.61837   5.4843255 3.502941e-04
#> ENSG00000142937 ENSG00000142937  0.05479432 14.85548   0.6336537 5.416308e-01
#> ENSG00000143158 ENSG00000143158 -0.11072705 11.39560  -0.8285994 4.281585e-01
#> ENSG00000147677 ENSG00000147677 -0.49046405 14.86806  -6.1253199 1.540396e-04
#> ENSG00000154723 ENSG00000154723  0.51827494 12.83740   3.7177988 4.550409e-03
#> ENSG00000161016 ENSG00000161016 -0.56354657 17.58235  -8.7982581 8.477064e-06
#> ENSG00000163956 ENSG00000163956  0.94051231 12.78486   6.4224059 1.072152e-04
#> ENSG00000165502 ENSG00000165502 -0.02780231 13.15988  -0.2915952 7.770192e-01
#> ENSG00000166337 ENSG00000166337  0.01432168 11.06957   0.1089877 9.155361e-01
#> ENSG00000166710 ENSG00000166710 -0.17917168 17.28166  -1.7479337 1.134513e-01
#> ENSG00000168264 ENSG00000168264 -2.60070945 13.47768 -25.8924141 5.847276e-10
#> ENSG00000175130 ENSG00000175130  2.93700441 12.15769  18.3043755 1.373050e-08
#> ENSG00000178741 ENSG00000178741 -0.94962267 12.75139  -3.4152442 7.366058e-03
#> ENSG00000182117 ENSG00000182117 -1.35082858 12.97909 -14.0893767 1.436218e-07
#>                    adj.P.Val          B
#> ENSG00000008018 7.108180e-04 -0.9762151
#> ENSG00000044574 3.164198e-05  2.5910685
#> ENSG00000065978 4.730425e-07  8.0111261
#> ENSG00000068697 5.038941e-02 -5.3615089
#> ENSG00000075624 5.095308e-10 16.3945961
#> ENSG00000100138 2.729487e-03 -2.6705184
#> ENSG00000101361 9.871326e-06  4.0250396
#> ENSG00000105193 1.949919e-02 -5.0438558
#> ENSG00000105887 5.095308e-10 16.6874669
#> ENSG00000105968 3.687726e-09 14.0228327
#> ENSG00000107223 8.551403e-06  4.7020176
#> ENSG00000107581 9.678490e-07  6.7765357
#> ENSG00000115758 9.063452e-05  1.2028423
#> ENSG00000119421 7.381355e-05  2.1518980
#> ENSG00000127884 2.204958e-01 -6.8155261
#> ENSG00000131495 4.065003e-03 -2.5718070
#> ENSG00000138279 8.897133e-01 -8.1306364
#> ENSG00000142541 6.387716e-04 -0.8589252
#> ENSG00000142937 5.996627e-01 -8.2830573
#> ENSG00000143158 4.915894e-01 -7.4699389
#> ENSG00000147677 2.984517e-04 -0.1956067
#> ENSG00000154723 6.717271e-03 -3.4253950
#> ENSG00000161016 2.388991e-05  2.7773208
#> ENSG00000163956 2.215780e-04  0.5962404
#> ENSG00000165502 8.306067e-01 -8.1284852
#> ENSG00000166337 9.155361e-01 -7.7909942
#> ENSG00000166710 1.406796e-01 -7.2160624
#> ENSG00000168264 4.531639e-09 13.5863478
#> ENSG00000175130 8.512911e-08 10.3730790
#> ENSG00000178741 1.037945e-02 -3.9119713
#> ENSG00000182117 6.360393e-07  7.7427974
#> 
#> $noiseq
#>                    BM_mean    JJ_mean            M            D       prob
#> ENSG00000008018  14764.121   7549.673  0.967609362 7.214448e+03 0.90322581
#> ENSG00000044574  20935.866  31948.426 -0.609768274 1.101256e+04 0.92473118
#> ENSG00000065978  14061.040  32529.547 -1.210047401 1.846851e+04 0.96774194
#> ENSG00000068697   3179.174   2368.093  0.424926259 8.110809e+02 0.44086022
#> ENSG00000075624  53126.804 254051.310 -2.257608085 2.009245e+05 1.00000000
#> ENSG00000100138  21117.847  17089.172  0.305380261 4.028675e+03 0.69354839
#> ENSG00000101361  25244.475  41192.464 -0.706412751 1.594799e+04 0.95161290
#> ENSG00000105193  38831.788  33466.887  0.214503793 5.364901e+03 0.63978495
#> ENSG00000105887  82893.643  12577.399  2.720427852 7.031624e+04 1.00000000
#> ENSG00000105968  50230.711   9373.394  2.421926302 4.085732e+04 1.00000000
#> ENSG00000107223   4510.232   9831.039 -1.124142153 5.320807e+03 0.85483871
#> ENSG00000107581  59736.457  29839.535  1.001386620 2.989692e+04 0.97849462
#> ENSG00000115758  31277.568  57724.524 -0.884056040 2.644696e+04 0.96774194
#> ENSG00000119421   4069.435   2072.632  0.973364171 1.996802e+03 0.67204301
#> ENSG00000127884   2514.800   3083.469 -0.294110997 5.686696e+02 0.32258065
#> ENSG00000131495   1819.229   2970.283 -0.707273169 1.151054e+03 0.52150538
#> ENSG00000138279   8518.429   8853.521 -0.055663917 3.350919e+02 0.12365591
#> ENSG00000142541  10629.982  14937.162 -0.490766854 4.307180e+03 0.76881720
#> ENSG00000142937  29442.741  30169.294 -0.035168991 7.265539e+02 0.16666667
#> ENSG00000143158   2815.790   2582.216  0.124930215 2.335743e+02 0.13440860
#> ENSG00000147677  35494.320  25155.778  0.496698364 1.033854e+04 0.88709677
#> ENSG00000154723   6268.223   8790.313 -0.487857922 2.522089e+03 0.69892473
#> ENSG00000161016 237890.341 160199.776  0.570424565 7.769057e+04 0.95698925
#> ENSG00000163956   5219.090   9784.980 -0.906770588 4.565890e+03 0.83333333
#> ENSG00000165502   9178.759   8992.186  0.029627157 1.865725e+02 0.05913978
#> ENSG00000166337   2160.659   2165.729 -0.003381174 5.069771e+00 0.00000000
#> ENSG00000166710 170766.046 153265.523  0.155987951 1.750052e+04 0.60752688
#> ENSG00000168264  28331.214   4708.256  2.589127701 2.362296e+04 0.97849462
#> ENSG00000175130   1671.911  12607.288 -2.914687705 1.093538e+04 0.95161290
#> ENSG00000178741  10418.538   5078.194  1.036765432 5.340344e+03 0.85483871
#> ENSG00000182117  12880.766   5041.935  1.353168851 7.838830e+03 0.90860215
#>                       ranking
#> ENSG00000008018  7.214448e+03
#> ENSG00000044574 -1.101256e+04
#> ENSG00000065978 -1.846851e+04
#> ENSG00000068697  8.110810e+02
#> ENSG00000075624 -2.009245e+05
#> ENSG00000100138  4.028675e+03
#> ENSG00000101361 -1.594799e+04
#> ENSG00000105193  5.364901e+03
#> ENSG00000105887  7.031624e+04
#> ENSG00000105968  4.085732e+04
#> ENSG00000107223 -5.320807e+03
#> ENSG00000107581  2.989692e+04
#> ENSG00000115758 -2.644696e+04
#> ENSG00000119421  1.996803e+03
#> ENSG00000127884 -5.686697e+02
#> ENSG00000131495 -1.151054e+03
#> ENSG00000138279 -3.350919e+02
#> ENSG00000142541 -4.307180e+03
#> ENSG00000142937 -7.265539e+02
#> ENSG00000143158  2.335743e+02
#> ENSG00000147677  1.033854e+04
#> ENSG00000154723 -2.522090e+03
#> ENSG00000161016  7.769057e+04
#> ENSG00000163956 -4.565890e+03
#> ENSG00000165502  1.865725e+02
#> ENSG00000166337 -5.069772e+00
#> ENSG00000166710  1.750052e+04
#> ENSG00000168264  2.362296e+04
#> ENSG00000175130 -1.093538e+04
#> ENSG00000178741  5.340344e+03
#> ENSG00000182117  7.838831e+03
#> 
#> $ebseq
#>                 DEFound
#> ENSG00000008018      DE
#> ENSG00000044574      DE
#> ENSG00000065978      DE
#> ENSG00000068697      EE
#> ENSG00000075624      DE
#> ENSG00000100138      DE
#> ENSG00000101361      DE
#> ENSG00000105193      EE
#> ENSG00000105887      DE
#> ENSG00000105968      DE
#> ENSG00000107223      DE
#> ENSG00000107581      DE
#> ENSG00000115758      DE
#> ENSG00000119421      DE
#> ENSG00000127884      EE
#> ENSG00000131495      DE
#> ENSG00000138279      EE
#> ENSG00000142541      DE
#> ENSG00000142937      DE
#> ENSG00000143158      EE
#> ENSG00000147677      DE
#> ENSG00000154723      DE
#> ENSG00000161016      DE
#> ENSG00000163956      DE
#> ENSG00000165502      EE
#> ENSG00000166337      EE
#> ENSG00000166710      EE
#> ENSG00000168264      DE
#> ENSG00000175130      DE
#> ENSG00000178741      DE
#> ENSG00000182117      DE
#> 
#> $deseq2
#>                   baseMean log2FoldChange      lfcSE        stat        pvalue
#> ENSG00000008018   6737.331    -0.87080384 0.13290473  -6.5520905  5.673713e-11
#> ENSG00000044574  16331.144     0.67214248 0.07546860   8.9062539  5.278595e-19
#> ENSG00000065978  14489.744     1.28209072 0.06875535  18.6471420  1.332104e-77
#> ENSG00000068697   1696.358    -0.34016420 0.14919311  -2.2800261  2.260614e-02
#> ENSG00000075624  96510.567     2.32351986 0.11131366  20.8736275  9.300167e-97
#> ENSG00000100138  11766.278    -0.24513719 0.07407922  -3.3091222  9.358898e-04
#> ENSG00000101361  20586.805     0.75505043 0.10134711   7.4501424  9.323957e-14
#> ENSG00000105193  22389.005    -0.14087576 0.10266606  -1.3721747  1.700091e-01
#> ENSG00000105887  29118.330    -2.67532007 0.11995046 -22.3035421 3.413811e-110
#> ENSG00000105968  18191.302    -2.38437339 0.10704468 -22.2745621 6.521205e-110
#> ENSG00000107223   4491.210     1.22111881 0.11418535  10.6941816  1.083721e-26
#> ENSG00000107581  27488.681    -0.95344872 0.11959572  -7.9722649  1.557924e-15
#> ENSG00000115758  27472.040     0.92256198 0.14381014   6.4151385  1.406949e-10
#> ENSG00000119421   1882.216    -0.91788078 0.14007783  -6.5526486  5.652544e-11
#> ENSG00000127884   1715.368     0.36409717 0.17853268   2.0393867  4.141145e-02
#> ENSG00000131495   1489.009     0.81268672 0.14142393   5.7464584  9.113203e-09
#> ENSG00000138279   5211.796     0.10652748 0.15507777   0.6869294  4.921272e-01
#> ENSG00000142541   7897.169     0.56033600 0.08668214   6.4642616  1.017946e-10
#> ENSG00000142937  18274.171     0.12482677 0.07317452   1.7058776  8.803087e-02
#> ENSG00000143158   1658.818    -0.03208471 0.11331781  -0.2831391  7.770702e-01
#> ENSG00000147677  18672.984    -0.42841583 0.09817804  -4.3636626  1.279028e-05
#> ENSG00000154723   4616.547     0.59499328 0.12584399   4.7280231  2.267164e-06
#> ENSG00000161016 123002.695    -0.50134808 0.10762556  -4.6582620  3.188902e-06
#> ENSG00000163956   4640.373     0.99944276 0.14430471   6.9259192  4.331536e-12
#> ENSG00000165502   5651.367     0.03406712 0.11530885   0.2954423  7.676560e-01
#> ENSG00000166337   1324.602     0.08105787 0.11226337   0.7220331  4.702741e-01
#> ENSG00000166710  98856.097    -0.10784763 0.13727109  -0.7856544  4.320700e-01
#> ENSG00000168264   9961.997    -2.53617028 0.09832161 -25.7946368 1.018494e-146
#> ENSG00000175130   4503.308     3.00756594 0.13389830  22.4615697 9.864118e-112
#> ENSG00000178741   4574.722    -0.91608013 0.30912670  -2.9634455  3.042159e-03
#> ENSG00000182117   5475.040    -1.28205189 0.10374193 -12.3580875  4.405127e-35
#>                          padj
#> ENSG00000008018  1.256322e-10
#> ENSG00000044574  1.818183e-18
#> ENSG00000065978  6.882537e-77
#> ENSG00000068697  3.046914e-02
#> ENSG00000075624  5.766104e-96
#> ENSG00000100138  1.381552e-03
#> ENSG00000101361  2.627660e-13
#> ENSG00000105193  2.027031e-01
#> ENSG00000105887 3.527604e-109
#> ENSG00000105968 5.053934e-109
#> ENSG00000107223  4.199421e-26
#> ENSG00000107581  4.829564e-15
#> ENSG00000115758  2.725964e-10
#> ENSG00000119421  1.256322e-10
#> ENSG00000127884  5.348979e-02
#> ENSG00000131495  1.661819e-08
#> ENSG00000138279  5.260670e-01
#> ENSG00000142541  2.103755e-10
#> ENSG00000142937  1.091583e-01
#> ENSG00000143158  7.770702e-01
#> ENSG00000147677  1.982494e-05
#> ENSG00000154723  3.904560e-06
#> ENSG00000161016  5.202946e-06
#> ENSG00000163956  1.118980e-11
#> ENSG00000165502  7.770702e-01
#> ENSG00000166337  5.206607e-01
#> ENSG00000166710  4.960804e-01
#> ENSG00000168264 3.157332e-145
#> ENSG00000175130 1.528938e-110
#> ENSG00000178741  4.286679e-03
#> ENSG00000182117  1.950842e-34
#> 
#> $samseq
#>                         Gene ID       Gene Name Score(d) Fold Change q-value(%)
#> ENSG00000008018 ENSG00000008018 ENSG00000008018     -4.5       0.552          0
#> ENSG00000044574 ENSG00000044574 ENSG00000044574      4.5        1.53          0
#> ENSG00000065978 ENSG00000065978 ENSG00000065978      4.5       2.414          0
#> ENSG00000068697 ENSG00000068697 ENSG00000068697     -3.1       0.696      2.688
#> ENSG00000075624 ENSG00000075624 ENSG00000075624      4.5       4.783          0
#> ENSG00000100138 ENSG00000100138 ENSG00000100138     -4.5        0.83          0
#> ENSG00000101361 ENSG00000101361 ENSG00000101361      4.5       1.633          0
#> ENSG00000105193 ENSG00000105193 ENSG00000105193     -3.6       0.828      1.403
#> ENSG00000105887 ENSG00000105887 ENSG00000105887     -4.5       0.149          0
#> ENSG00000105968 ENSG00000105968 ENSG00000105968     -4.5       0.185          0
#> ENSG00000107223 ENSG00000107223 ENSG00000107223      4.5       2.154          0
#> ENSG00000107581 ENSG00000107581 ENSG00000107581     -4.5        0.48          0
#> ENSG00000115758 ENSG00000115758 ENSG00000115758      4.5       1.621          0
#> ENSG00000119421 ENSG00000119421 ENSG00000119421     -4.5       0.497          0
#> ENSG00000127884 ENSG00000127884 ENSG00000127884      2.6       1.353      7.444
#> ENSG00000131495 ENSG00000131495 ENSG00000131495      4.5       1.866          0
#> ENSG00000138279            <NA>            <NA>     <NA>        <NA>       <NA>
#> ENSG00000142541 ENSG00000142541 ENSG00000142541      4.5       1.385          0
#> ENSG00000142937            <NA>            <NA>     <NA>        <NA>       <NA>
#> ENSG00000143158 ENSG00000143158 ENSG00000143158     -1.4       1.006     15.532
#> ENSG00000147677 ENSG00000147677 ENSG00000147677     -4.5       0.647          0
#> ENSG00000154723 ENSG00000154723 ENSG00000154723      4.5         1.4          0
#> ENSG00000161016 ENSG00000161016 ENSG00000161016     -4.5       0.642          0
#> ENSG00000163956 ENSG00000163956 ENSG00000163956      4.5       2.047          0
#> ENSG00000165502 ENSG00000165502 ENSG00000165502    -0.45       1.036     18.433
#> ENSG00000166337            <NA>            <NA>     <NA>        <NA>       <NA>
#> ENSG00000166710 ENSG00000166710 ENSG00000166710    -2.65       0.854      2.688
#> ENSG00000168264 ENSG00000168264 ENSG00000168264     -4.5       0.156          0
#> ENSG00000175130 ENSG00000175130 ENSG00000175130      4.5       8.553          0
#> ENSG00000178741 ENSG00000178741 ENSG00000178741    -4.45       0.432          0
#> ENSG00000182117 ENSG00000182117 ENSG00000182117     -4.5       0.404          0
#>