Execute DESeq2 DGE default analysis is configured to read only kallisto files, and count data

Execute DESeq2 DGE default analysis is configured to read only kallisto files, and count data

runDeseq2(
  countMatrix,
  groupName,
  numberReplics,
  designExperiment,
  fitTypeParam = "local",
  pathDirRuns = ".",
  pathReportFile = "report_txi.txt",
  subDirRuns = "dir_runs",
  fileKallisto = "abundance.tsv"
)

Arguments

countMatrix

original table count or abundance data

groupName

text, name of samples or treatment

numberReplics

number of replicate (technical or biological) by sample

designExperiment

replicate and treatment by samples

fitTypeParam

either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity

pathDirRuns

parent directory that contains one folder with all kallisto output

pathReportFile

file path with report by sample, conditions and runs

subDirRuns

sub dir by pathDirRun

fileKallisto

file output kallisto, exemaple: "abundance.tsv" or "quant.sf.gz"

Value

extracted results from DESeq2 (generated by DESeq2::results function)

Examples

groupNameModel = c("BM","JJ")
numberReplicsModel = 3
toolResult <- NULL
toolResult$deseq2 <- runDeseq2(countMatrix = gse95077,
                               groupName = groupNameModel,
                               numberReplics = numberReplicsModel,
                               fitTypeParam = "local")
#> [1] "Dataset is COUNT data"
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing