frameAllGenes

This function creates a data.frame for each result of the diffrential expression method performed. It was created to ensure that all genes been listed, even thought the gene does not appear as diffrentially expressed, in this case NA data is inserted. This functis is executed into runExpression() function, so there is no need to call it separatelly.

Usage

frameAllGenes(cons_list, countMatrix)

Arguments

cons_list

A named list of data.frames, each corresponding to a method (e.g., edgeR, DESeq2, etc.)

countMatrix

RData object with table count, where line name is gene name, and name column is treat name

Value

list with analysis expression containing: all methods executed and their results, ensuring thar all genes has been listed;

Examples

set.seed(42)
counts <- matrix(
  as.integer(c(
    rnbinom(200, mu = 10,  size = 1), 
    rnbinom(200, mu = 100, size = 1) 
  )),
  nrow = 100,
  dimnames = list(paste0("gene", seq_len(100)),
                    paste0("sample", seq_len(4)))
)
groups_info <- c("control", "control", "treated", "treated")
treats <- c("control", "treated")
cons_result <- runExpression(
  numberReplics = 2,
  groupName = treats,
  rDataFrameCount = counts,
  experimentName = "test_cons",
  controlDeseq2 = "control",
  contrastDeseq2 = "treated",
  outDirPath = tempdir(),
  printResults = FALSE
)
#> calcNormFactors has been renamed to normLibSizes
#> Using classic mode.
#> baySeq libsizes
#> baySeq getPriors
#> Finding priors...
#> Warning: The '@replicates' slot is not a factor; converting now.
#> done.
#> Finding posterior likelihoods...
#> Length of priorReps:0
#> Length of priorSubset:100
#> Length of subset:100
#> Length of postRows:100
#> Analysing part 1 of 1
#> Preparing data...
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> .
#> done.
#> Estimating likelihoods...
#> ...done!
#> .
#> done.
#> Available columns in results_raw:
#> Warning: ‘>=’ not meaningful for factors
#> calcNormFactors has been renamed to normLibSizes
#> Removing intercept from test coefficients
#> [1] "Computing (M,D) values..."
#> [1] "Computing probability of differential expression..."
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Initial number of DE patterns = 2
#> Final number of DE patterns = 2
#> converting counts to integer mode
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> Estimating sequencing depths...
#> Resampling to get new data matrices...
#> perm= 1
#> perm= 2
#> perm= 3
#> perm= 4
#> perm= 5
#> perm= 6
#> perm= 7
#> perm= 8
#> perm= 9
#> perm= 10
#> perm= 11
#> perm= 12
#> perm= 13
#> perm= 14
#> perm= 15
#> perm= 16
#> perm= 17
#> perm= 18
#> perm= 19
#> perm= 20
#> perm= 21
#> perm= 22
#> perm= 23
#> perm= 24
#> Number of thresholds chosen (all possible thresholds) = 11
#> Getting all the cutoffs for the thresholds...
#> Getting number of false positives in the permutation...
#> Differential expression analysis completed.

genes_list <- results(cons_result)
framed <- frameAllGenes(cons_list = genes_list, countMatrix = counts)
#>  - 1edger
#>  - 2limma
#>  - 3noiseq
#>  - 4ebseq
#>  - 5deseq2
#>  - 6samseq
head(framed)
#> $edger
#>               logFC   logCPM     PValue       FDR
#> gene1    1.91851893 14.31983 0.30380003 0.9229003
#> gene10   0.35586486 12.76616 0.85978097 0.9844509
#> gene100 -0.45899561 12.50677 0.80778278 0.9844509
#> gene11  -0.47097526 12.15632 0.80640618 0.9844509
#> gene12   1.75471303 12.92653 0.37049742 0.9262436
#> gene13  -0.86464875 13.87887 0.63510838 0.9844509
#> gene14  -2.58521650 12.96298 0.18170524 0.9229003
#> gene15   1.46037191 12.58686 0.44972332 0.9844509
#> gene16  -2.44417586 13.54679 0.19918576 0.9229003
#> gene17   0.64015812 13.48271 0.73033888 0.9844509
#> gene18   7.56628308 11.79115 0.04691513 0.9229003
#> gene19  -1.32957236 12.54407 0.49706196 0.9844509
#> gene2    2.64240624 14.01498 0.16927854 0.9229003
#> gene20   1.55637643 14.96470 0.39169885 0.9553631
#> gene21   1.97303729 14.03424 0.29005233 0.9229003
#> gene22  -0.23274518 13.06343 0.91597016 0.9844509
#> gene23  -1.35847083 14.11221 0.45314954 0.9844509
#> gene24  -0.71644208 13.63902 0.69485896 0.9844509
#> gene25  -0.23666614 12.27686 0.92319277 0.9844509
#> gene26  -0.23272220 13.54688 0.90025079 0.9844509
#> gene27   0.72107649 14.04800 0.68685301 0.9844509
#> gene28   0.61905103 12.86427 0.75556632 0.9844509
#> gene29   1.81469998 13.12676 0.34490558 0.9229003
#> gene3    2.50605094 13.81737 0.19822521 0.9229003
#> gene30   1.79767352 13.43587 0.35070213 0.9229003
#> gene31   1.90006002 13.75315 0.31671228 0.9229003
#> gene32  -0.00209869 14.02423 1.00000000 1.0000000
#> gene33   4.77690337 14.20141 0.02350611 0.9229003
#> gene34   0.41689707 13.10660 0.83578463 0.9844509
#> gene35  -1.83309066 12.25116 0.34751380 0.9229003
#> gene36  -1.18497147 12.93477 0.52058700 0.9844509
#> gene37   1.25768063 12.99963 0.52352398 0.9844509
#> gene38  -2.02036434 12.92290 0.28258492 0.9229003
#> gene39   0.32328251 13.51549 0.86741633 0.9844509
#> gene4   -0.70072845 13.44741 0.70533871 0.9844509
#> gene40   1.12467085 14.12337 0.54621607 0.9844509
#> gene41  -0.40226974 13.21640 0.83739351 0.9844509
#> gene42  -0.46142628 13.69552 0.79482897 0.9844509
#> gene43   3.51944600 11.77107 0.12694342 0.9229003
#> gene44  -0.12849869 12.73884 0.96384857 0.9844509
#> gene45   2.28592110 12.94262 0.25911853 0.9229003
#> gene46   1.89928682 13.66637 0.31665615 0.9229003
#> gene47  -1.79361786 12.76514 0.33081996 0.9229003
#> gene48   4.97129015 13.13236 0.03934096 0.9229003
#> gene49  -4.15722131 14.35469 0.04016486 0.9229003
#> gene5    4.23410645 13.39541 0.05800870 0.9229003
#> gene50   1.33925366 13.54486 0.46582840 0.9844509
#> gene51  -1.34399986 12.66763 0.48086084 0.9844509
#> gene52  -0.58049348 12.97026 0.76111361 0.9844509
#> gene53   0.96476149 13.03007 0.61534683 0.9844509
#> gene54   2.17619202 13.56808 0.25761440 0.9229003
#> gene55   0.13060025 13.72693 0.94732110 0.9844509
#> gene56   3.21594681 13.18187 0.11759732 0.9229003
#> gene57   0.64552224 13.35528 0.73853360 0.9844509
#> gene58   3.94458571 13.52841 0.05975553 0.9229003
#> gene59  -0.20006465 11.77365 0.96276489 0.9844509
#> gene6    0.01444237 12.80693 1.00000000 1.0000000
#> gene60  -0.50335184 12.91123 0.81131453 0.9844509
#> gene61  -0.71319834 13.30165 0.69730347 0.9844509
#> gene62   1.04007825 14.05359 0.56636741 0.9844509
#> gene63   2.13691415 13.50166 0.26031424 0.9229003
#> gene64   0.30197834 13.21004 0.88502497 0.9844509
#> gene65   0.43247632 13.81197 0.82664017 0.9844509
#> gene66   2.63128465 13.84502 0.17945366 0.9229003
#> gene67   2.46904625 13.24518 0.20999098 0.9229003
#> gene68   0.58432157 13.32695 0.75679976 0.9844509
#> gene69  -0.86304054 12.98814 0.65259691 0.9844509
#> gene7   -1.18671619 13.55805 0.52164621 0.9844509
#> gene70   0.85970428 14.32714 0.63680589 0.9844509
#> gene71   0.39035680 12.90316 0.85648544 0.9844509
#> gene72   0.21061914 12.44094 0.95490693 0.9844509
#> gene73   3.49027342 13.29511 0.09178734 0.9229003
#> gene74   2.02193844 13.12498 0.30055195 0.9229003
#> gene75   2.22548120 14.48019 0.24099190 0.9229003
#> gene76   1.16722391 13.40312 0.53584031 0.9844509
#> gene77  -2.04221854 11.12406 0.33452799 0.9229003
#> gene78  -0.93580964 13.41966 0.60431639 0.9844509
#> gene79   2.65138043 13.17749 0.18849783 0.9229003
#> gene8   -0.60250023 13.26622 0.74802457 0.9844509
#> gene80   1.16873029 12.51736 0.53319659 0.9844509
#> gene81  -2.05365184 12.67172 0.28023894 0.9229003
#> gene82   2.94400789 13.73446 0.14259923 0.9229003
#> gene83   0.34510941 14.14968 0.85706461 0.9844509
#> gene84   0.34840351 13.35831 0.86399241 0.9844509
#> gene85   0.44497222 12.16711 0.86143205 0.9844509
#> gene86  -1.26452292 13.08689 0.49435174 0.9844509
#> gene87   0.37938253 13.91298 0.84551525 0.9844509
#> gene88   0.19086614 11.83684 0.96476187 0.9844509
#> gene89   1.23436924 13.70665 0.50429136 0.9844509
#> gene9   -0.27792138 13.52638 0.87852474 0.9844509
#> gene90   1.97726414 13.87053 0.29504986 0.9229003
#> gene91  -0.24029387 13.42545 0.90208055 0.9844509
#> gene92   0.67874780 13.10734 0.72773693 0.9844509
#> gene93  -1.67672917 13.11452 0.36222692 0.9262436
#> gene94   3.88253383 12.87727 0.06793397 0.9229003
#> gene95   0.74866318 12.37760 0.69436271 0.9844509
#> gene96  -3.49918326 13.22683 0.08108431 0.9229003
#> gene97   2.23114955 13.48341 0.25044419 0.9229003
#> gene98  -1.29268082 13.39973 0.48937572 0.9844509
#> gene99  -2.27399743 12.76408 0.22315130 0.9229003
#> 
#> $limma
#>              ID        logFC  AveExpr            t    P.Value adj.P.Val
#> gene1     gene1  2.743449248 12.99149  0.934429524 0.35120918 0.9722839
#> gene10   gene10 -1.517534437 10.56262 -0.419782549 0.67509490 0.9722839
#> gene100 gene100 -1.541674497 12.00503 -0.582075704 0.56117125 0.9722839
#> gene11   gene11 -1.359884890 11.78552 -0.482890021 0.62970198 0.9722839
#> gene12   gene12  1.028796327 12.60167  0.373690725 0.70903025 0.9722839
#> gene13   gene13 -3.020475443 12.94528 -1.713471658 0.08817528 0.9722839
#> gene14   gene14 -3.568694250 12.19447 -1.524565044 0.12894733 0.9722839
#> gene15   gene15  0.671044653 12.06239  0.229054534 0.81906059 0.9722839
#> gene16   gene16 -2.530035119 12.86505 -1.390706041 0.16586001 0.9722839
#> gene17   gene17 -1.868451869 11.72398 -0.663009722 0.50808773 0.9722839
#> gene18   gene18  2.537936431 10.39760  0.831174109 0.40686647 0.9722839
#> gene19   gene19 -0.477253436 11.57791 -0.155763818 0.87637629 0.9722839
#> gene2     gene2 -2.090668045 11.53230 -0.707509396 0.48007359 0.9722839
#> gene20   gene20  0.772798130 14.03996  0.397785435 0.69121261 0.9722839
#> gene21   gene21  1.307878470 13.05022  0.499875309 0.61771240 0.9722839
#> gene22   gene22 -1.026800178 12.58397 -0.437314899 0.66235483 0.9722839
#> gene23   gene23 -1.392316622 13.62571 -0.921830760 0.35772752 0.9722839
#> gene24   gene24 -1.005325308 13.38428 -0.571171081 0.56852494 0.9722839
#> gene25   gene25 -0.891496091 12.19334 -0.341644580 0.73297708 0.9722839
#> gene26   gene26 -1.076553837 13.36591 -0.629013145 0.53005823 0.9722839
#> gene27   gene27  1.842741537 13.05346  0.669340863 0.50405001 0.9722839
#> gene28   gene28 -2.889574224 11.13988 -0.875387086 0.38241340 0.9722839
#> gene29   gene29  0.934261810 12.73818  0.350554816 0.72629130 0.9722839
#> gene3     gene3  3.083646822 12.68782  1.048507597 0.29567020 0.9722839
#> gene30   gene30  1.278988648 13.04020  0.490237090 0.62450351 0.9722839
#> gene31   gene31  0.650726459 12.39535  0.236364027 0.81339210 0.9722839
#> gene32   gene32 -0.512753049 13.90203 -0.324558991 0.74585408 0.9722839
#> gene33   gene33  3.560018495 12.28595  1.210484421 0.22752166 0.9722839
#> gene34   gene34  0.611558784 12.55300  0.229469807 0.81873829 0.9722839
#> gene35   gene35 -2.784144707 11.78315 -1.003589674 0.31678910 0.9722839
#> gene36   gene36 -2.177298657 12.61114 -1.040906720 0.29917551 0.9722839
#> gene37   gene37 -1.982107890 11.62828 -0.688873591 0.49170117 0.9722839
#> gene38   gene38 -2.386274394 12.49482 -1.115625027 0.26592133 0.9722839
#> gene39   gene39  0.042910983 13.26987  0.019417418 0.98452747 0.9980659
#> gene4     gene4 -1.765867207 13.09299 -1.002974662 0.31708501 0.9722839
#> gene40   gene40  0.614170184 13.94901  0.323121786 0.74694059 0.9722839
#> gene41   gene41  0.930118016 12.21135  0.321387162 0.74825263 0.9722839
#> gene42   gene42 -0.966313347 13.51024 -0.584272589 0.55969538 0.9722839
#> gene43   gene43  2.206669160 10.99792  0.736805912 0.46210381 0.9722839
#> gene44   gene44 -0.440540553 12.47965 -0.175988060 0.86048139 0.9722839
#> gene45   gene45  0.135273235 11.59973  0.044363688 0.96465877 0.9980659
#> gene46   gene46  0.529019770 12.89823  0.212700466 0.83177724 0.9722839
#> gene47   gene47 -2.451784651 12.37593 -1.092883344 0.27575966 0.9722839
#> gene48   gene48  3.633689353 11.71159  1.235534129 0.21808110 0.9722839
#> gene49   gene49 -4.148558994 12.74570 -2.145111029 0.03314977 0.9722839
#> gene5     gene5  3.594109727 12.06029  1.222076188 0.22311717 0.9722839
#> gene50   gene50  1.779465775 12.51862  0.606092990 0.54514027 0.9722839
#> gene51   gene51 -0.493208783 11.62295 -0.162712178 0.87090934 0.9722839
#> gene52   gene52 -1.092105776 12.77848 -0.498312205 0.61881155 0.9722839
#> gene53   gene53 -0.005934635 12.73392 -0.002427050 0.99806591 0.9980659
#> gene54   gene54  1.878645755 12.97491  0.674218070 0.50095121 0.9722839
#> gene55   gene55  0.026586371 13.37857  0.012730626 0.98985539 0.9980659
#> gene56   gene56  2.582103538 11.97828  0.877971873 0.38101247 0.9722839
#> gene57   gene57  1.052779208 12.05879  0.357364563 0.72119571 0.9722839
#> gene58   gene58  2.927638952 12.51953  0.995461497 0.32071471 0.9722839
#> gene59   gene59 -0.731445374 10.81976 -0.219966143 0.82612188 0.9722839
#> gene6     gene6 -0.285978423 12.55787 -0.115187909 0.90841175 0.9767868
#> gene60   gene60 -0.436791302 12.36638 -0.169744049 0.86538300 0.9722839
#> gene61   gene61 -1.909444395 12.92270 -1.023826518 0.30715398 0.9722839
#> gene62   gene62  0.408192469 13.52139  0.188063636 0.85101744 0.9722839
#> gene63   gene63  2.555861573 12.57636  0.869049029 0.38586202 0.9722839
#> gene64   gene64  0.197408967 12.76996  0.080149993 0.93619810 0.9959554
#> gene65   gene65  0.930172335 13.07403  0.373989973 0.70880795 0.9722839
#> gene66   gene66  2.199031678 13.20207  0.801879891 0.42357417 0.9722839
#> gene67   gene67  1.726820563 12.77207  0.608805435 0.54334428 0.9722839
#> gene68   gene68 -2.977843203 11.65782 -1.024317566 0.30692263 0.9722839
#> gene69   gene69 -0.950453048 12.55911 -0.401174070 0.68872026 0.9722839
#> gene7     gene7 -2.045067678 12.82518 -1.086599748 0.27852159 0.9722839
#> gene70   gene70  0.213518191 14.21652  0.132779603 0.89450118 0.9722839
#> gene71   gene71  1.270777314 12.05958  0.432092177 0.66613991 0.9722839
#> gene72   gene72 -0.586888524 12.25420 -0.224130246 0.82288475 0.9722839
#> gene73   gene73  3.003713468 12.00780  1.021328502 0.30833264 0.9722839
#> gene74   gene74  1.030886567 12.60271  0.374449965 0.70846629 0.9722839
#> gene75   gene75  1.837205427 13.91421  0.791947336 0.42932967 0.9722839
#> gene76   gene76  0.470426176 13.23337  0.202278932 0.83990428 0.9722839
#> gene77   gene77 -3.889867348 10.05451 -0.987533364 0.32457443 0.9722839
#> gene78   gene78 -1.859339149 12.91521 -0.992638355 0.32208565 0.9722839
#> gene79   gene79  0.988803722 12.15962  0.337909281 0.73578597 0.9722839
#> gene8     gene8  0.807016134 12.21882  0.281547089 0.77858191 0.9722839
#> gene80   gene80  1.390254106 11.82627  0.471424492 0.63785147 0.9722839
#> gene81   gene81 -1.376224810 11.25557 -0.431712796 0.66641519 0.9722839
#> gene82   gene82  2.310976594 12.77812  0.787127650 0.43213891 0.9722839
#> gene83   gene83 -0.225962827 14.03656 -0.142546209 0.88679208 0.9722839
#> gene84   gene84  1.535741974 12.42079  0.524817468 0.60029133 0.9722839
#> gene85   gene85 -0.956186522 11.61794 -0.322063640 0.74774087 0.9722839
#> gene86   gene86 -2.754269863 12.22860 -1.187287135 0.23652261 0.9722839
#> gene87   gene87  1.874297939 12.83854  0.663803432 0.50758060 0.9722839
#> gene88   gene88 -1.250863066 11.04357 -0.375145655 0.70794967 0.9722839
#> gene89   gene89  1.340734612 12.44548  0.461020068 0.64528515 0.9722839
#> gene9     gene9 -1.656524882 13.03072 -0.924520881 0.35632930 0.9722839
#> gene90   gene90  1.232352516 13.49556  0.517770189 0.60519086 0.9722839
#> gene91   gene91 -0.544632212 12.88478 -0.240699288 0.81003475 0.9722839
#> gene92   gene92 -0.078048665 12.80175 -0.032500207 0.97410556 0.9980659
#> gene93   gene93 -2.064897145 12.73728 -1.043296490 0.29807041 0.9722839
#> gene94   gene94  1.854833021 11.24069  0.622322330 0.53443879 0.9722839
#> gene95   gene95  0.011542825 12.23292  0.004236625 0.99662389 0.9980659
#> gene96   gene96 -2.129072105 11.31196 -0.675424108 0.50018651 0.9722839
#> gene97   gene97  1.517263198 12.98831  0.565547289 0.57233548 0.9722839
#> gene98   gene98 -1.654731716 12.90063 -0.848124424 0.39738256 0.9722839
#> gene99   gene99 -2.636448800 12.23363 -1.132821852 0.25864529 0.9722839
#>                 B
#> gene1   -4.595194
#> gene10  -4.595435
#> gene100 -4.595591
#> gene11  -4.595603
#> gene12  -4.595687
#> gene13  -4.592017
#> gene14  -4.593914
#> gene15  -4.595672
#> gene16  -4.593715
#> gene17  -4.595473
#> gene18  -4.595286
#> gene19  -4.595639
#> gene2   -4.595406
#> gene20  -4.596234
#> gene21  -4.595667
#> gene22  -4.595853
#> gene23  -4.595452
#> gene24  -4.596206
#> gene25  -4.595768
#> gene26  -4.596151
#> gene27  -4.595484
#> gene28  -4.595227
#> gene29  -4.595737
#> gene3   -4.595063
#> gene30  -4.595678
#> gene31  -4.595742
#> gene32  -4.596909
#> gene33  -4.594852
#> gene34  -4.595786
#> gene35  -4.595116
#> gene36  -4.595026
#> gene37  -4.595437
#> gene38  -4.594854
#> gene39  -4.596142
#> gene4   -4.595113
#> gene40  -4.596358
#> gene41  -4.595655
#> gene42  -4.596323
#> gene43  -4.595375
#> gene44  -4.595892
#> gene45  -4.595656
#> gene46  -4.595891
#> gene47  -4.594928
#> gene48  -4.594816
#> gene49  -4.590320
#> gene5   -4.594835
#> gene50  -4.595487
#> gene51  -4.595649
#> gene52  -4.595902
#> gene53  -4.595955
#> gene54  -4.595471
#> gene55  -4.596265
#> gene56  -4.595253
#> gene57  -4.595622
#> gene58  -4.595125
#> gene59  -4.595550
#> gene6   -4.595920
#> gene60  -4.595853
#> gene61  -4.595053
#> gene62  -4.596143
#> gene63  -4.595262
#> gene64  -4.595938
#> gene65  -4.595815
#> gene66  -4.595357
#> gene67  -4.595511
#> gene68  -4.595091
#> gene69  -4.595867
#> gene7   -4.594867
#> gene70  -4.597016
#> gene71  -4.595590
#> gene72  -4.595812
#> gene73  -4.595095
#> gene74  -4.595687
#> gene75  -4.595466
#> gene76  -4.596006
#> gene77  -4.595128
#> gene78  -4.595143
#> gene79  -4.595637
#> gene8   -4.595680
#> gene80  -4.595567
#> gene81  -4.595520
#> gene82  -4.595341
#> gene83  -4.597065
#> gene84  -4.595543
#> gene85  -4.595628
#> gene86  -4.594741
#> gene87  -4.595471
#> gene88  -4.595506
#> gene89  -4.595585
#> gene9   -4.595348
#> gene90  -4.595765
#> gene91  -4.596039
#> gene92  -4.595985
#> gene93  -4.595009
#> gene94  -4.595465
#> gene95  -4.595793
#> gene96  -4.595393
#> gene97  -4.595592
#> gene98  -4.595489
#> gene99  -4.594860
#> 
#> $noiseq
#>         control_mean treated_mean            M           D  prob      ranking
#> gene1      8791.2088   27177.0556 -1.628255708 18385.84684 0.465 -18385.84691
#> gene10     6043.9560    6476.2345 -0.099662065   432.27849 0.010   -432.27850
#> gene100    7142.8571    3989.8231  0.840176500  3153.03408 0.160   3153.03419
#> gene11     5494.5055    2717.7056  1.015600482  2776.79993 0.160   2776.80012
#> gene12     3846.1538    9367.4107 -1.284233842  5521.25682 0.260  -5521.25697
#> gene13    21978.0220    7979.6461  1.461664877 13998.37586 0.445  13998.37593
#> gene14    15934.0659    1908.1762  3.061848115 14025.88969 0.625  14025.89002
#> gene15     3846.1538    7112.2933 -0.886898344  3266.13943 0.190  -3266.13956
#> gene16    21978.0220    3006.8232  2.869749616 18971.19880 0.665  18971.19902
#> gene17    11538.4615   10176.9400  0.181147043  1361.52156 0.050   1361.52157
#> gene18      274.7253    4452.4112 -4.018525302  4177.68597 0.320  -4177.68790
#> gene19    10989.0110    2602.0585  2.078336237  8386.95247 0.395   8386.95273
#> gene2      5494.5055   20932.1152 -1.929656553 15437.60969 0.515 -15437.60981
#> gene20    21428.5714   36370.9957 -0.763252749 14942.42429 0.215 -14942.42431
#> gene21     8791.2088   20064.7623 -1.190530614 11273.55355 0.340 -11273.55362
#> gene22    11538.4615    6534.0581  0.820399699  5004.40348 0.185   5004.40355
#> gene23    27472.5275    7863.9991  1.804654587 19608.52840 0.490  19608.52848
#> gene24    20329.6703    7748.3520  1.391625414 12581.31830 0.410  12581.31838
#> gene25     6043.9560    3469.4114  0.800802262  2574.54469 0.125   2574.54481
#> gene26    14835.1648    8384.4108  0.823239651  6450.75406 0.190   6450.75411
#> gene27    17032.9670   17231.4097 -0.016710971   198.44270 0.005   -198.44271
#> gene28     7142.8571    6418.4110  0.154285091   724.44613 0.030    724.44615
#> gene29     4395.6044   11159.9399 -1.344195798  6764.33547 0.300  -6764.33560
#> gene3      4395.6044   19602.1742 -2.156880224 15206.56977 0.545 -15206.56992
#> gene30     6043.9560   13704.1749 -1.181050391  7660.21881 0.300  -7660.21891
#> gene31     6043.9560   15959.2922 -1.400831599  9915.33620 0.375  -9915.33630
#> gene32    18681.3187   13761.9984  0.440906317  4919.32030 0.100   4919.32032
#> gene33     1648.3516   27177.0556 -4.043293207 25528.70398 0.780 -25528.70430
#> gene34     7692.3077    7632.7050  0.011222042    59.60271 0.005     59.60271
#> gene35     8791.2088    1561.2351  2.493373737  7229.97368 0.405   7229.97411
#> gene36    12637.3626    3585.0584  1.817626885  9052.30424 0.385   9052.30442
#> gene37     5494.5055    8210.9402 -0.579557786  2716.43472 0.120  -2716.43478
#> gene38    13736.2637    2428.5879  2.499800006 11307.67579 0.520  11307.67606
#> gene39    13186.8132   10466.0576  0.333377853  2720.75559 0.085   2720.75562
#> gene4     17032.9670    6418.4110  1.408041683 10614.55602 0.380  10614.55612
#> gene40    12637.3626   19891.2918 -0.654441560  7253.92914 0.155  -7253.92917
#> gene41    10989.0110    6591.8816  0.737299320  4397.12941 0.180   4397.12947
#> gene42    17032.9670    8673.5284  0.973638859  8359.43864 0.270   8359.43870
#> gene43      549.4505    4336.7642 -2.980557452  3787.31365 0.290  -3787.31482
#> gene44     8791.2088    5146.2935  0.772527808  3644.91528 0.175   3644.91536
#> gene45     2747.2527    9193.9401 -1.742693622  6446.68735 0.325  -6446.68758
#> gene46     6043.9560   17231.4097 -1.511475663 11187.45369 0.400 -11187.45380
#> gene47    12087.9121    2370.7644  2.350140853  9717.14766 0.460   9717.14795
#> gene48      549.4505   12778.9985 -4.539641321 12229.54795 0.640 -12229.54879
#> gene49    42857.1429    1734.7057  4.626772862 41122.43718 0.875  41122.43744
#> gene5      1098.9011   15381.0570 -3.807021197 14282.15592 0.660 -14282.15642
#> gene50     8791.2088   12894.6455 -0.552638661  4103.43675 0.115  -4103.43679
#> gene51    12087.9121    2948.9997  2.035267516  9138.91243 0.415   9138.91266
#> gene52    10439.5604    4914.9994  1.086797816  5524.56102 0.245   5524.56112
#> gene53     6593.4066    8153.1167 -0.306327613  1559.71009 0.065  -1559.71012
#> gene54     5494.5055   15554.5276 -1.501273029 10060.02209 0.390 -10060.02220
#> gene55    13736.2637   10986.4693  0.322261820  2749.79444 0.085   2749.79446
#> gene56     2197.8022   13588.5278 -2.628255708 11390.72562 0.540 -11390.72592
#> gene57     8241.7582   10697.3517 -0.376229626  2455.59344 0.075  -2455.59347
#> gene58     1648.3516   16826.6451 -3.351651603 15178.29343 0.655 -15178.29380
#> gene59     4395.6044    2197.2939  1.000333725  2198.31054 0.130   2198.31076
#> gene6      8791.2088    5666.7052  0.633551395  3124.50358 0.130   3124.50364
#> gene60    11538.4615    5146.2935  1.164845231  6392.16803 0.260   6392.16813
#> gene61    14285.7143    5724.5287  1.319344337  8561.18555 0.340   8561.18565
#> gene62    14285.7143   19313.0566 -0.435003335  5027.34227 0.100  -5027.34228
#> gene63     5494.5055   14744.9983 -1.424164103  9250.49277 0.365  -9250.49288
#> gene64     8791.2088    7921.8226  0.150229156   869.38619 0.030    869.38621
#> gene65    12637.3626   13704.1749 -0.116920054  1066.81222 0.025  -1066.81223
#> gene66     4395.6044   20006.9388 -2.186366989 15611.33443 0.555 -15611.33458
#> gene67     3296.7033   12778.9985 -1.954678820  9482.29520 0.410  -9482.29540
#> gene68     9890.1099   10466.0576 -0.081659647   575.94770 0.010   -575.94771
#> gene69    10989.0110    4683.7053  1.230339331  6305.30566 0.270   6305.30578
#> gene7     18131.8681    6302.7640  1.524471033 11829.10417 0.415  11829.10427
#> gene70    17582.4176   21857.2915 -0.313981185  4274.87396 0.085  -4274.87398
#> gene71     6593.4066    6823.1757 -0.049419310   229.76907 0.010   -229.76908
#> gene72     5494.5055    4336.7642  0.341370643  1157.74130 0.060   1157.74135
#> gene73     1648.3516   14860.6453 -3.172400810 13212.29366 0.600 -13212.29404
#> gene74     3846.1538   11564.7045 -1.588240029  7718.55068 0.355  -7718.55084
#> gene75     8791.2088   29605.6436 -1.751738761 20814.43478 0.490 -20814.43486
#> gene76     7692.3077   12027.2927 -0.644823557  4334.98501 0.145  -4334.98506
#> gene77     3846.1538     636.0587  2.596184542  3210.09510 0.235   3210.09615
#> gene78    18131.8681    6360.5875  1.511295645 11771.28064 0.410  11771.28074
#> gene79     2747.2527   13125.9396 -2.256359154 10378.68688 0.470 -10378.68713
#> gene8     12087.9121    6129.2934  0.979772403  5958.61869 0.235   5958.61877
#> gene80     4395.6044    6244.9404 -0.506626263  1849.33605 0.075  -1849.33612
#> gene81    13736.2637    2081.6468  2.722192427 11654.61692 0.545  11654.61724
#> gene82     3296.7033   19833.4683 -2.588841027 16536.76496 0.605 -16536.76516
#> gene83    18131.8681   15959.2922  0.184130901  2172.57589 0.050   2172.57590
#> gene84     9340.6593    9367.4107 -0.004125923    26.75132 0.000    -26.75132
#> gene85     4395.6044    4163.2936  0.078336237   232.31077 0.010    232.31078
#> gene86    15384.6154    4336.7642  1.826797470 11047.85119 0.425  11047.85134
#> gene87    13736.2637   14224.5866 -0.050397077   488.32283 0.010   -488.32283
#> gene88     3296.7033    2602.0585  0.341370643   694.64478 0.040    694.64486
#> gene89    10439.5604   15843.6452 -0.601843331  5404.08476 0.145  -5404.08479
#> gene9     14835.1648    9078.2930  0.708527992  5756.87179 0.170   5756.87183
#> gene90     7142.8571   18445.7037 -1.368711657 11302.84657 0.395 -11302.84665
#> gene91    15384.6154    7632.7050  1.011222042  7751.91040 0.270   7751.91047
#> gene92     8241.7582    8846.9990 -0.102236008   605.24072 0.015   -605.24073
#> gene93    14835.1648    3295.9408  2.170258727 11539.22405 0.490  11539.22425
#> gene94     1098.9011    9887.8224 -3.169591276  8788.92127 0.470  -8788.92184
#> gene95     4395.6044    5088.4700 -0.211170380   692.86559 0.030   -692.86563
#> gene96    23626.3736    1214.2940  4.282208571 22412.07965 0.785  22412.08006
#> gene97     4395.6044   15207.5864 -1.790657750 10811.98205 0.420 -10811.98220
#> gene98    19780.2198    5204.1170  1.926333144 14576.10275 0.505  14576.10287
#> gene99    12637.3626    1908.1762  2.727429076 10729.18639 0.520  10729.18674
#> 
#> $ebseq
#>         DEFound
#> gene1        EE
#> gene10       EE
#> gene100      EE
#> gene11       EE
#> gene12       EE
#> gene13       EE
#> gene14       DE
#> gene15       EE
#> gene16       DE
#> gene17       EE
#> gene18       EE
#> gene19       EE
#> gene2        EE
#> gene20       EE
#> gene21       EE
#> gene22       EE
#> gene23       EE
#> gene24       EE
#> gene25       EE
#> gene26       EE
#> gene27       EE
#> gene28       EE
#> gene29       EE
#> gene3        DE
#> gene30       EE
#> gene31       EE
#> gene32       EE
#> gene33       EE
#> gene34       EE
#> gene35       DE
#> gene36       EE
#> gene37       EE
#> gene38       DE
#> gene39       EE
#> gene4        EE
#> gene40       EE
#> gene41       EE
#> gene42       EE
#> gene43       EE
#> gene44       EE
#> gene45       EE
#> gene46       EE
#> gene47       DE
#> gene48       EE
#> gene49       DE
#> gene5        EE
#> gene50       EE
#> gene51       EE
#> gene52       EE
#> gene53       EE
#> gene54       EE
#> gene55       EE
#> gene56       EE
#> gene57       EE
#> gene58       EE
#> gene59       EE
#> gene6        EE
#> gene60       EE
#> gene61       EE
#> gene62       EE
#> gene63       EE
#> gene64       EE
#> gene65       EE
#> gene66       DE
#> gene67       EE
#> gene68       EE
#> gene69       EE
#> gene7        EE
#> gene70       EE
#> gene71       EE
#> gene72       EE
#> gene73       EE
#> gene74       EE
#> gene75       EE
#> gene76       EE
#> gene77       EE
#> gene78       EE
#> gene79       EE
#> gene8        EE
#> gene80       EE
#> gene81       EE
#> gene82       EE
#> gene83       EE
#> gene84       EE
#> gene85       EE
#> gene86       EE
#> gene87       EE
#> gene88       EE
#> gene89       EE
#> gene9        EE
#> gene90       EE
#> gene91       EE
#> gene92       EE
#> gene93       DE
#> gene94       EE
#> gene95       EE
#> gene96       DE
#> gene97       EE
#> gene98       EE
#> gene99       DE
#> 
#> $deseq2
#>          baseMean log2FoldChange     lfcSE         stat       pvalue       padj
#> gene1   46.759801    1.490934434 1.5690041  0.950242532 0.3419890318 0.69004757
#> gene10  16.142876   -0.144192139 2.4584107 -0.058652584 0.9532288264 0.99863514
#> gene100 14.272479   -0.932048099 1.4858920 -0.627265053 0.5304855073 0.86964837
#> gene11  11.495617   -0.761858274 1.4832821 -0.513630067 0.6075106541 0.88045022
#> gene12  17.430391    1.359643241 1.2263517  1.108689464 0.2675641686 0.63705754
#> gene13  41.590296   -1.121355868 1.5366610 -0.729735380 0.4655519502 0.83983990
#> gene14  22.347439   -3.063946039 1.1829948 -2.589991172 0.0095978393 0.17994440
#> gene15  13.375534    0.985013283 1.4854637  0.663101569 0.5072655057 0.84544251
#> gene16  33.808828   -2.819060805 1.2075702 -2.334490211 0.0195700681 0.23374500
#> gene17  29.805717    0.363468615 2.0085143  0.180963916 0.8563958992 0.99863514
#> gene18   7.039212    4.794583505 2.0017671  2.395175473 0.0166124129 0.23374500
#> gene19  15.944525   -1.726483546 1.7309529 -0.997417973 0.3185616707 0.67926078
#> gene2   41.241057    2.387763397 2.0713792  1.152740828 0.2490167781 0.62983345
#> gene20  82.608659    1.256649302 1.7696463  0.710113249 0.4776339108 0.83983990
#> gene21  41.906903    1.667739136 1.4806306  1.126370849 0.2600085805 0.63416727
#> gene22  21.376864   -0.732009554 1.3304843 -0.550182764 0.5821940246 0.88045022
#> gene23  49.002782   -1.677843740 1.1905053 -1.409354265 0.1587304403 0.55736189
#> gene24  34.227090   -1.123930225 1.1902441 -0.944285479 0.3450237843 0.69004757
#> gene25  12.092691   -0.646650093 1.2369034 -0.522797579 0.6011151242 0.88045022
#> gene26  31.679936   -0.539363775 1.2193886 -0.442323123 0.6582554009 0.88831195
#> gene27  42.171391    0.285437170 1.5587707  0.183116841 0.8547063422 0.99863514
#> gene28  18.888521    0.352150489 2.0826868  0.169084705 0.8657300157 0.99863514
#> gene29  19.907372    1.384509030 1.1914234  1.162063024 0.2452098655 0.62983345
#> gene3   32.961919    2.145488464 1.5329364  1.399593947 0.1616349477 0.55736189
#> gene30  25.570843    1.371398649 1.1683967  1.173744073 0.2404975512 0.62983345
#> gene31  34.428100    1.659503607 1.7321213  0.958075862 0.3380245061 0.69004757
#> gene32  42.551030   -0.409782770 0.9986571 -0.410333789 0.6815611092 0.88831195
#> gene33  45.383617    4.498825403 1.6485204  2.729008036 0.0063525156 0.17994440
#> gene34  22.057970    0.114090962 1.3895690  0.082105285 0.9345629909 0.99863514
#> gene35  13.166245   -2.167086344 1.3790259 -1.571461722 0.1160754389 0.49713309
#> gene36  21.248679   -1.498497252 1.2055714 -1.242976783 0.2138763889 0.62983345
#> gene37  20.320508    1.001347925 1.8094367  0.553403137 0.5799873916 0.88045022
#> gene38  21.445704   -2.392981279 1.0807475 -2.214190878 0.0268156511 0.26815651
#> gene39  29.387053   -0.102111124 1.2448214 -0.082028735 0.9346238634 0.99863514
#> gene4   30.334511   -1.010896314 1.3376501 -0.755725526 0.4498137767 0.83983990
#> gene40  43.384317    0.739493152 0.9596633  0.770575607 0.4409585243 0.83983990
#> gene41  24.366690   -0.784853683 1.6398110 -0.478624470 0.6322058069 0.88831195
#> gene42  35.326083   -0.778904860 1.1595874 -0.671708643 0.5017691981 0.84544251
#> gene43   6.715599    3.166738701 1.7216649  1.839346681 0.0658642111 0.38026279
#> gene44  16.991764   -0.567252267 1.3073858 -0.433882854 0.6643735109 0.88831195
#> gene45  18.783555    2.048974783 1.7884574  1.145665991 0.2519333808 0.62983345
#> gene46  28.781572    1.425786044 1.4315193  0.995995016 0.3192525666 0.67926078
#> gene47  19.115454   -2.126586457 1.1918245 -1.784311699 0.0743730302 0.38026279
#> gene48  20.080762    4.787228094 1.8593895  2.574623646 0.0100349212 0.17994440
#> gene49  61.339544   -4.474585636 1.3586095 -3.293503889 0.0009894697 0.09894697
#> gene5   24.362800    3.983785846 1.5627519  2.549212021 0.0107966639 0.17994440
#> gene50  29.711687    1.011814411 1.7102082  0.591632299 0.5540968407 0.88045022
#> gene51  17.300041   -1.808074106 1.7202151 -1.051074435 0.2932244044 0.66641910
#> gene52  20.998896   -0.905452488 1.1499411 -0.787390334 0.4310534035 0.83983990
#> gene53  20.620828    0.653070240 1.2627839  0.517167048 0.6050395527 0.88045022
#> gene54  28.300499    1.782689581 1.3001014  1.371192716 0.1703148884 0.56771629
#> gene55  35.076912   -0.178516599 1.2939410 -0.137963480 0.8902692801 0.99863514
#> gene56  19.458190    2.739250998 1.5645229  1.750853866 0.0799710853 0.38081469
#> gene57  24.626426    0.182369833 1.7722090  0.102905375 0.9180380714 0.99863514
#> gene58  26.983642    3.625938250 1.4116846  2.568518738 0.0102134190 0.17994440
#> gene59   8.413024   -0.506185787 1.7645655 -0.286861435 0.7742184203 0.96777303
#> gene6   17.743055   -0.418994267 1.2840041 -0.326318479 0.7441834086 0.94200431
#> gene60  19.690572   -0.968198214 1.4071833 -0.688039862 0.4914276796 0.84544251
#> gene61  27.244516   -0.999381682 1.3459015 -0.742537002 0.4577620417 0.83983990
#> gene62  40.518705    0.550303306 1.2805232  0.429748810 0.6673783733 0.88831195
#> gene63  27.018280    1.744077669 1.5046946  1.159090782 0.2464191829 0.62983345
#> gene64  24.003083    0.004198939 1.3407197  0.003131854 0.9975011459 0.99863514
#> gene65  35.296465    0.016635461 1.4824723  0.011221431 0.9910467813 0.99863514
#> gene66  33.939843    2.289698261 1.1993682  1.909086968 0.0562508736 0.38026279
#> gene67  21.836046    2.099089926 1.1220444  1.870772536 0.0613766158 0.38026279
#> gene68  24.148584    0.062186595 1.9290992  0.032236080 0.9742837837 0.99863514
#> gene69  21.106216   -1.271311259 1.2608571 -1.008291338 0.3133146201 0.67926078
#> gene7   32.032588   -1.661266111 1.4437923 -1.150626817 0.2498857958 0.62983345
#> gene70  50.682206    0.438613862 0.9115113  0.481194104 0.6303785507 0.88831195
#> gene71  18.508518    0.011682632 1.6383545  0.007130711 0.9943105643 0.99863514
#> gene72  13.655211   -0.101208839 1.3527064 -0.074819517 0.9403583132 0.99863514
#> gene73  21.165181    3.100341012 1.5214914  2.037698708 0.0415800735 0.34650061
#> gene74  19.526257    1.583463381 1.2767883  1.240192632 0.2149041529 0.62983345
#> gene75  55.275438    1.910883650 1.1205486  1.705310891 0.0881364574 0.40062026
#> gene76  25.580243    0.753024397 1.0630219  0.708380912 0.4787087431 0.83983990
#> gene77   5.350468   -2.574265432 1.3425401 -1.917458880 0.0551796601 0.38026279
#> gene78  28.961531   -1.434711552 1.3611469 -1.054046043 0.2918618375 0.66641910
#> gene79  19.405465    2.181615606 1.5235751  1.431905566 0.1521708491 0.55736189
#> gene8   25.835469   -0.938015507 1.6984836 -0.552266450 0.5807658130 0.88045022
#> gene80  13.158742    0.727722169 1.6333852  0.445530030 0.6559367628 0.88831195
#> gene81  17.861884   -2.561708651 1.8070273 -1.417636913 0.1562967975 0.55736189
#> gene82  29.654513    2.528586615 1.3991584  1.807219669 0.0707280353 0.38026279
#> gene83  47.040830    0.018917205 1.0102684  0.018724930 0.9850605407 0.99863514
#> gene84  25.980455   -0.020903974 1.6441814 -0.012713910 0.9898560410 0.99863514
#> gene85  10.258370   -0.054293480 1.5391337 -0.035275349 0.9718601794 0.99863514
#> gene86  23.110890   -1.785726566 1.4616643 -1.221707752 0.2218181639 0.62983345
#> gene87  38.499691   -0.002821424 1.6493784 -0.001710598 0.9986351409 0.99863514
#> gene88   8.701260   -0.064652223 1.6789707 -0.038507060 0.9692834028 0.99863514
#> gene89  30.949045    0.725647226 1.7828706  0.407010599 0.6840002038 0.88831195
#> gene9   29.746432   -0.767752162 1.3875108 -0.553330604 0.5800370487 0.88045022
#> gene90  36.145914    1.640344599 1.1241950  1.459128219 0.1445298207 0.55736189
#> gene91  29.262043   -0.570527240 1.4775495 -0.386130710 0.6993998666 0.89666650
#> gene92  20.880034    0.205728924 1.2147327  0.169361483 0.8655123182 0.99863514
#> gene93  24.400279   -2.016734642 1.1367880 -1.774064040 0.0760525580 0.38026279
#> gene94  17.312246    3.644704772 1.9871413  1.834144764 0.0666325081 0.38026279
#> gene95  12.144769    0.328765740 1.2792238  0.257004089 0.7971756168 0.98416743
#> gene96  28.025703   -3.907765140 1.7969946 -2.174611486 0.0296592429 0.26962948
#> gene97  25.641088    1.827454209 1.1731987  1.557668123 0.1193119407 0.49713309
#> gene98  29.325298   -1.764820770 1.3238347 -1.333112671 0.1824948316 0.58869301
#> gene99  19.227835   -2.676404002 1.1599628 -2.307318870 0.0210370503 0.23374500
#> 
#> $samseq
#>         Gene ID Gene Name Score(d)    Fold Change q-value(%)
#> gene1      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene10     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene100    <NA>      <NA>     <NA>           <NA>       <NA>
#> gene11     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene12   gene12    gene12     1.95          2.883     64.286
#> gene13     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene14     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene15     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene16     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene17     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene18   gene18    gene18        2 2331119974.565     64.286
#> gene19     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene2      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene20     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene21     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene22     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene23     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene24     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene25     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene26     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene27     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene28     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene29   gene29    gene29      1.8          2.973     64.286
#> gene3     gene3     gene3        2          5.088     64.286
#> gene30   gene30    gene30        2          2.831     64.286
#> gene31     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene32     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene33   gene33    gene33     1.95         23.148     64.286
#> gene34     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene35     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene36     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene37     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene38     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene39     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene4      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene40   gene40    gene40     1.75          1.874     64.286
#> gene41     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene42     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene43   gene43    gene43        2         10.301     64.286
#> gene44     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene45     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene46     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene47     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene48   gene48    gene48        2         28.057     64.286
#> gene49     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene5     gene5     gene5        2         16.712     64.286
#> gene50     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene51     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene52     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene53     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene54   gene54    gene54        2          3.661     64.286
#> gene55     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene56   gene56    gene56      1.8          7.582     64.286
#> gene57     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene58   gene58    gene58        2         13.174     64.286
#> gene59     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene6      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene60     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene61     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene62     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene63   gene63    gene63        2          3.517     64.286
#> gene64     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene65     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene66   gene66    gene66        2          5.421     64.286
#> gene67   gene67    gene67        2          4.717     64.286
#> gene68     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene69     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene7      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene70     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene71     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene72     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene73   gene73    gene73        2          9.889     64.286
#> gene74     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene75   gene75    gene75        2          4.117     64.286
#> gene76     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene77     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene78     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene79     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene8      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene80     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene81     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene82   gene82    gene82      1.9          6.728     64.286
#> gene83     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene84     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene85     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene86     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene87     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene88     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene89     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene9      <NA>      <NA>     <NA>           <NA>       <NA>
#> gene90   gene90    gene90        2          3.312     64.286
#> gene91     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene92     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene93     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene94     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene95     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene96     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene97   gene97    gene97        2          4.031     64.286
#> gene98     <NA>      <NA>     <NA>           <NA>       <NA>
#> gene99     <NA>      <NA>     <NA>           <NA>       <NA>
#>