frameAllGenes

This function creates a data.frame for each result of the diffrential expression method performed. It was created to ensure that all genes been listed, even thought the gene does not appear as diffrentially expressed, in this case NA data is inserted. This functis is executed into runExpression() function, so there is no need to call it separatelly.

Usage

frameAllGenes(cons_list, countMatrix)

Arguments

cons_list

list generated by runExpression() function execution

countMatrix

RData object with table count, where line name is gene name, and name column is treat name

Value

list with analysis expression containing: all methods executed and their results, ensuring thar all genes has been listed;

Examples

data(gse95077)
treats = c("BM", "JJ")
cons_result <- runExpression(numberReplics = 3,
                             groupName = treats,
                              rDataFrameCount = gse95077,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              controlDeseq2 = "BM",
                              contrastDeseq2 = "JJ",
                              outDirPath = "." )
#> Using classic mode.
#> Getting annotation of the Homo Sapiens...
#> Using reference genome 38.
#> Calculating gene expression values...
#> RQ fit ......
#> SQN 
#> 
#>  
#>  ===== ERROR: KnowSeq execution is failed === 
#>  Error in mclustBIC(data = structure(c(-3.52446659876917, -2.42801002792396, : could not find function "mclustBIC"
#>  
#> Removing intercept from test coefficients
#> 
#>  ------------ limma executed!
#> [1] "Computing (M,D) values..."
#> [1] "Computing probability of differential expression..."
#> 
#>  ------------ NOISeq executed! 
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Initial number of DE patterns = 2
#> Final number of DE patterns = 2
#> 
#>  ------------ EBSeq executed!
#> Warning: some variables in design formula are characters, converting to factors
#> [1] "Dataset is COUNT data"
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> 
#>  ------------ DESeq2 executed!
#> Estimating sequencing depths...
#> Resampling to get new data matrices...
#> perm= 1
#> perm= 2
#> perm= 3
#> perm= 4
#> perm= 5
#> perm= 6
#> perm= 7
#> perm= 8
#> perm= 9
#> perm= 10
#> perm= 11
#> perm= 12
#> perm= 13
#> perm= 14
#> perm= 15
#> perm= 16
#> perm= 17
#> perm= 18
#> perm= 19
#> perm= 20
#> perm= 21
#> perm= 22
#> perm= 23
#> perm= 24
#> perm= 25
#> perm= 26
#> perm= 27
#> perm= 28
#> perm= 29
#> perm= 30
#> perm= 31
#> perm= 32
#> perm= 33
#> perm= 34
#> perm= 35
#> perm= 36
#> perm= 37
#> perm= 38
#> perm= 39
#> perm= 40
#> perm= 41
#> perm= 42
#> perm= 43
#> perm= 44
#> perm= 45
#> perm= 46
#> perm= 47
#> perm= 48
#> perm= 49
#> perm= 50
#> perm= 51
#> perm= 52
#> perm= 53
#> perm= 54
#> perm= 55
#> perm= 56
#> perm= 57
#> perm= 58
#> perm= 59
#> perm= 60
#> perm= 61
#> perm= 62
#> perm= 63
#> perm= 64
#> perm= 65
#> perm= 66
#> perm= 67
#> perm= 68
#> perm= 69
#> perm= 70
#> perm= 71
#> perm= 72
#> perm= 73
#> perm= 74
#> perm= 75
#> perm= 76
#> perm= 77
#> perm= 78
#> perm= 79
#> perm= 80
#> perm= 81
#> perm= 82
#> perm= 83
#> perm= 84
#> perm= 85
#> perm= 86
#> perm= 87
#> perm= 88
#> perm= 89
#> perm= 90
#> perm= 91
#> perm= 92
#> perm= 93
#> perm= 94
#> perm= 95
#> perm= 96
#> perm= 97
#> perm= 98
#> perm= 99
#> perm= 100
#> Number of thresholds chosen (all possible thresholds) = 8
#> Getting all the cutoffs for the thresholds...
#> Getting number of false positives in the permutation...
#> 
#>  ------------ SAMSeq executed!
#> 
#>  ------------ DIFERENTIAL EXPRESSION ANALYSIS WAS COMPLETE!
cons_rseult2 <- frameAllGenes(cons_list = cons_result,
                              countMatrix=gse95077)
#>  -  1 edger -  2 limma -  3 noiseq -  4 ebseq -  5 deseq2 -  6 samseq