Execute DESeq2 DGE default analysis is configured to read only kallisto files, and count data

Execute DESeq2 DGE default analysis is configured to read only kallisto files, and count data

Usage

runDeseq2(
  countMatrix,
  groupName,
  numberReplics,
  controlGroup,
  contrastGroup,
  fitTypeParam = "local"
)

Arguments

countMatrix

original table count or abundance data

groupName

list of text, name of samples or treatment

numberReplics

number of replicate (technical or biological) by sample

controlGroup

specific reference level used as control (this name should be in groupName List)

contrastGroup

specific reference level that you need compare (this name should be in groupName List)

fitTypeParam

either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity

Value

extracted results from DESeq2 (generated by DESeq2::results function)

Examples

groupNameModel = c("BM","JJ")
numberReplicsModel = 3
toolResult <- NULL
data(gse95077)
toolResult$deseq2 <- runDeseq2(countMatrix = gse95077,
                               groupName = groupNameModel,
                               numberReplics = numberReplicsModel,
                               controlGroup = "BM",
                               contrastGroup = "JJ",
                               fitTypeParam = "local")
#> Warning: some variables in design formula are characters, converting to factors
#> [1] "Dataset is COUNT data"
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing