Make expression analysis of 7 tools, and return a list of results by each tool.

Make expression analysis of 7 tools, and return a list of results by each tool.

Usage

runExpression(
  numberReplics,
  groupName,
  tableCountPath = "data/gse95077.csv",
  sepCharacter = ",",
  rDataFrameCount = NULL,
  experimentName = "genericExperiment",
  outDirPath = "../../consexpression2_results/",
  printResults = FALSE,
  fitTypeDeseq2 = "local",
  controlDeseq2 = "",
  contrastDeseq2 = "",
  methodNormLimma = "TMM",
  methodAdjPvalueLimma = "BH",
  numberTopTableLimma = 1e+06,
  filterIdKnowseq = "ensembl_gene_id",
  notSapiensKnowseq = FALSE,
  methodNormEdgeR = "TMM",
  normNoiseq = "rpkm",
  kNoiseq = 0.5,
  factorNoiseq = "Tissue",
  lcNoiseq = 0,
  replicatesNoiseq = "technical",
  condExpNoiseq = c(""),
  respTypeSamseq = "Two class unpaired",
  npermSamseq = 100,
  fdrEbseq = 0.05,
  maxRoundEbseq = 50,
  methodDeResultsEbseq = "robust",
  deNovoAanalysis = FALSE,
  progressShiny = NULL
)

Arguments

numberReplics

number of replicate (technical or biological) by sample

groupName

text, name of samples or treatment

tableCountPath

path to csv file that contains count data or abundance data (local)

sepCharacter

character used to split csv data, can be comma or tab

rDataFrameCount

RData object with table count, where line name is gene name, and name column is treat name

experimentName

text, name of experiment

outDirPath

path to write output, need be a directory (local)

printResults

logical variable: TRUE print report by each tool, FALSE print only consensus result

fitTypeDeseq2

either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity

controlDeseq2

group of samples that represents control in experiment, used by DESeq2; Default is "".

contrastDeseq2

specific reference level that you need compare (this name should be in groupName List)

methodNormLimma

normalization method to be used (limma)

methodAdjPvalueLimma

correction method, a character string. Can be abbreviated (limma)

numberTopTableLimma

maximum number of genes to list (limma)

filterIdKnowseq

The attribute used as filter to return the rest of the attributes.

notSapiensKnowseq

A boolean value that indicates if the user wants the human annotation or another annotation available in BiomaRt. The possible not human dataset can be consulted by calling the following function: biomaRt::listDatasets(useMart("ensembl")).

methodNormEdgeR

normalization method to be used in edgeR::calcNormFactors(), default: "TMM"

normNoiseq

Normalization method t can be one of "rpkm" (default), "uqua" (upper quartile), "tmm" (trimmed mean of M) or "n" (no normalization). Default: "rpkm".

kNoiseq

Counts equal to 0 are replaced by k. By default, k = 0.5

factorNoiseq

A string indicating the name of factor whose levels are the conditions to be compared.

lcNoiseq

Length correction is done by dividing expression by length^lc. By default, lc = 0

replicatesNoiseq

In this argument, the type of replicates to be used is defined: "technical", "biological" or "no" replicates. By default, "technical" replicates option is chosen.

condExpNoiseq

A vector containing the two conditions to be compared by the differential expression algorithm (needed when the factor contains more than 2 different conditions).

respTypeSamseq

Problem type: "Quantitative" for a continuous parameter; "Two class unpaired" for two classes with unpaired observations; "Survival" for censored survival outcome; "Multiclass": more than 2 groups; "Two class paired" for two classes with paired observations. Default: "Two class unpaired".

npermSamseq

Number of permutations used to estimate false discovery rates. Default 100

fdrEbseq

parameter used in EBTest function: fdr False Discovery Rate cutt off

maxRoundEbseq

parameter used in EBTest function: Number of iterations. The default value is 50.

methodDeResultsEbseq

parameter used in GetDEResults function: "robust" or "classic". Using the "robust" option, EBSeq is more robust to genes with outliers and genes with extremely small variances. Using the "classic" option, the results will be more comparable to those obtained by using the GetPPMat() function from earlier version (<= 1.7.0) of EBSeq

deNovoAanalysis

boolean value (TRUE if dataset don`t have a reference genome)

progressShiny

shiny app element, used to show execution progress

Value

list with all analysis expression

Examples

data(gse95077)
treats = c("BM", "JJ")
cons_result <- runExpression(numberReplics = 3,
                              groupName = treats,
                              rDataFrameCount = gse95077,
                              sepCharacter = ",",
                              experimentName = "test_cons",
                              controlDeseq2 = "BM",
                              contrastDeseq2 = "JJ",
                              outDirPath = "." )
#> Using classic mode.
#> Getting annotation of the Homo Sapiens...
#> Using reference genome 38.
#> Calculating gene expression values...
#> RQ fit ......
#> SQN 
#> 
#>  
#>  ===== ERROR: KnowSeq execution is failed === 
#>  Error in mclustBIC(data = structure(c(-3.52446659876917, -2.42801002792396, : could not find function "mclustBIC"
#>  
#> Removing intercept from test coefficients
#> 
#>  ------------ limma executed!
#> [1] "Computing (M,D) values..."
#> [1] "Computing probability of differential expression..."
#> 
#>  ------------ NOISeq executed! 
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Warning: `expect_is()` was deprecated in the 3rd edition.
#> ℹ Use `expect_type()`, `expect_s3_class()`, or `expect_s4_class()` instead
#> Initial number of DE patterns = 2
#> Final number of DE patterns = 2
#> 
#>  ------------ EBSeq executed!
#> Warning: some variables in design formula are characters, converting to factors
#> [1] "Dataset is COUNT data"
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
#> 
#>  ------------ DESeq2 executed!
#> Estimating sequencing depths...
#> Resampling to get new data matrices...
#> perm= 1
#> perm= 2
#> perm= 3
#> perm= 4
#> perm= 5
#> perm= 6
#> perm= 7
#> perm= 8
#> perm= 9
#> perm= 10
#> perm= 11
#> perm= 12
#> perm= 13
#> perm= 14
#> perm= 15
#> perm= 16
#> perm= 17
#> perm= 18
#> perm= 19
#> perm= 20
#> perm= 21
#> perm= 22
#> perm= 23
#> perm= 24
#> perm= 25
#> perm= 26
#> perm= 27
#> perm= 28
#> perm= 29
#> perm= 30
#> perm= 31
#> perm= 32
#> perm= 33
#> perm= 34
#> perm= 35
#> perm= 36
#> perm= 37
#> perm= 38
#> perm= 39
#> perm= 40
#> perm= 41
#> perm= 42
#> perm= 43
#> perm= 44
#> perm= 45
#> perm= 46
#> perm= 47
#> perm= 48
#> perm= 49
#> perm= 50
#> perm= 51
#> perm= 52
#> perm= 53
#> perm= 54
#> perm= 55
#> perm= 56
#> perm= 57
#> perm= 58
#> perm= 59
#> perm= 60
#> perm= 61
#> perm= 62
#> perm= 63
#> perm= 64
#> perm= 65
#> perm= 66
#> perm= 67
#> perm= 68
#> perm= 69
#> perm= 70
#> perm= 71
#> perm= 72
#> perm= 73
#> perm= 74
#> perm= 75
#> perm= 76
#> perm= 77
#> perm= 78
#> perm= 79
#> perm= 80
#> perm= 81
#> perm= 82
#> perm= 83
#> perm= 84
#> perm= 85
#> perm= 86
#> perm= 87
#> perm= 88
#> perm= 89
#> perm= 90
#> perm= 91
#> perm= 92
#> perm= 93
#> perm= 94
#> perm= 95
#> perm= 96
#> perm= 97
#> perm= 98
#> perm= 99
#> perm= 100
#> Number of thresholds chosen (all possible thresholds) = 7
#> Getting all the cutoffs for the thresholds...
#> Getting number of false positives in the permutation...
#> 
#>  ------------ SAMSeq executed!
#> 
#>  ------------ DIFERENTIAL EXPRESSION ANALYSIS WAS COMPLETE!
expDef_result <- expressionDefinition(resultTool = cons_result, groups = treats)